apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
motor==3.4.0
beanie==1.26.0
pymongo==4.6.3
pydantic==2.7.0
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
cloudinary==1.40.0
Pillow==10.3.0
imagehash==4.3.1
numpy==1.26.4
groq==0.9.0
python-dotenv==1.0.1
stripe==9.9.0
redis==5.0.4
slowapi==0.1.9
structlog==24.2.0
sentry-sdk[fastapi]==2.5.1
pytest==8.2.0
pytest-asyncio==0.23.7